removed some compile warning if compile with VS.NET
git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@684 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/windows/snprintf.c
@@ -547,7 +547,7 @@ static long round (LDOUBLE value)
 {
   long intpart;
 
-  intpart = value;
+  intpart = (long) value;
   value = value - intpart;
   if (value >= 0.5)
     intpart++;
@@ -592,7 +592,7 @@ static void fmtfp (char *buffer, size_t *currlen, size_t maxlen,
   if (flags & DP_F_UP) caps = 1; /* Should characters be upper case? */
 #endif
 
-  intpart = ufvalue;
+  intpart = (long) ufvalue;
 
   /* 
    * Sorry, we only support 9 digits past the decimal because of our 
@@ -609,7 +609,7 @@ static void fmtfp (char *buffer, size_t *currlen, size_t maxlen,
   if (fracpart >= pow10 (max))
   {
     intpart++;
-    fracpart -= pow10 (max);
+    fracpart -= (long) pow10 (max);
   }
 
 #ifdef DEBUG_SNPRINTF
